1. Move AppList to ui 2. Use AppList in AppListScreen and SearchResultScreen
