+++ core/ui/src/main/java/com/merxury/blocker/core/ui/applist/AppList.kt
@@ -1,0 +1,71 @@
+/*
+ * Copyright 2023 Blocker
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.merxury.blocker.core.ui.applist
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.ui.Modifier
+import com.merxury.blocker.core.ui.applist.model.AppItem
+
+@Composable
+fun AppList(
+    appList: List<AppItem>,
+    onAppItemClick: (String) -> Unit,
+    onClearCacheClick: (String) -> Unit,
+    onClearDataClick: (String) -> Unit,
+    onForceStopClick: (String) -> Unit,
+    onUninstallClick: (String) -> Unit,
+    onEnableClick: (String) -> Unit,
+    onDisableClick: (String) -> Unit,
+    onServiceStateUpdate: (String, Int) -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    val listState = rememberLazyListState()
+    Box(modifier) {
+        LazyColumn(
+            modifier = modifier,
+            state = listState,
+        ) {
+            itemsIndexed(appList, key = { _, item -> item.packageName }) { index, item ->
+                AppListItem(
+                    label = item.label,
+                    packageName = item.packageName,
+                    versionName = item.versionName,
+                    versionCode = item.versionCode,
+                    packageInfo = item.packageInfo,
+                    isAppEnabled = item.isEnabled,
+                    isAppRunning = item.isRunning,
+                    appServiceStatus = item.appServiceStatus,
+                    onClick = onAppItemClick,
+                    onClearCacheClick = onClearCacheClick,
+                    onClearDataClick = onClearDataClick,
+                    onForceStopClick = onForceStopClick,
+                    onUninstallClick = onUninstallClick,
+                    onEnableClick = onEnableClick,
+                    onDisableClick = onDisableClick,
+                )
+                LaunchedEffect(true) {
+                    onServiceStateUpdate(item.packageName, index)
+                }
+            }
+        }
+    }
+}
