+++ feature/applist/src/main/java/com/merxury/blocker/feature/applist/AppListScreen.kt
@@ -17,7 +17,6 @@
 package com.merxury.blocker.feature.applist
 
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.WindowInsets
@@ -28,14 +27,9 @@ import androidx.compose.foundation.layout.only
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.safeDrawing
 import androidx.compose.foundation.layout.windowInsetsPadding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.State
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
@@ -46,7 +40,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.merxury.blocker.core.designsystem.component.BlockerErrorAlertDialog
 import com.merxury.blocker.core.designsystem.component.BlockerTopAppBar
 import com.merxury.blocker.core.model.preference.AppSorting
-import com.merxury.blocker.core.ui.applist.AppListItem
+import com.merxury.blocker.core.ui.applist.AppList
 import com.merxury.blocker.core.ui.applist.model.AppItem
 import com.merxury.blocker.core.ui.screen.ErrorScreen
 import com.merxury.blocker.core.ui.screen.LoadingScreen
@@ -67,7 +61,7 @@ fun AppListRoute(
     val appList = viewModel.appListFlow.collectAsState()
     AppListScreen(
         uiState = uiState,
-        appList = appList,
+        appList = appList.value,
         onAppItemClick = navigateToAppDetail,
         onClearCacheClick = viewModel::clearCache,
         onClearDataClick = viewModel::clearData,
@@ -94,7 +88,7 @@ fun AppListRoute(
 @Composable
 fun AppListScreen(
     uiState: AppListUiState,
-    appList: State<List<AppItem>>,
+    appList: List<AppItem>,
     onAppItemClick: (String) -> Unit,
     onClearCacheClick: (String) -> Unit,
     onClearDataClick: (String) -> Unit,
@@ -141,7 +135,7 @@ fun AppListScreen(
                 }
 
                 is AppListUiState.Success -> {
-                    AppListContent(
+                    AppList(
                         appList = appList,
                         onAppItemClick = onAppItemClick,
                         onClearCacheClick = onClearCacheClick,
@@ -156,51 +150,6 @@ fun AppListScreen(
                 }
 
                 is AppListUiState.Error -> ErrorScreen(uiState.error)
-            }
-        }
-    }
-}
-
-@Composable
-fun AppListContent(
-    appList: State<List<AppItem>>,
-    onAppItemClick: (String) -> Unit,
-    onClearCacheClick: (String) -> Unit,
-    onClearDataClick: (String) -> Unit,
-    onForceStopClick: (String) -> Unit,
-    onUninstallClick: (String) -> Unit,
-    onEnableClick: (String) -> Unit,
-    onDisableClick: (String) -> Unit,
-    onServiceStateUpdate: (String, Int) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val listState = rememberLazyListState()
-    Box(modifier) {
-        LazyColumn(
-            modifier = modifier,
-            state = listState,
-        ) {
-            itemsIndexed(appList.value, key = { _, item -> item.packageName }) { index, item ->
-                AppListItem(
-                    label = item.label,
-                    packageName = item.packageName,
-                    versionName = item.versionName,
-                    versionCode = item.versionCode,
-                    packageInfo = item.packageInfo,
-                    isAppEnabled = item.isEnabled,
-                    isAppRunning = item.isRunning,
-                    appServiceStatus = item.appServiceStatus,
-                    onClick = onAppItemClick,
-                    onClearCacheClick = onClearCacheClick,
-                    onClearDataClick = onClearDataClick,
-                    onForceStopClick = onForceStopClick,
-                    onUninstallClick = onUninstallClick,
-                    onEnableClick = onEnableClick,
-                    onDisableClick = onDisableClick,
-                )
-                LaunchedEffect(true) {
-                    onServiceStateUpdate(item.packageName, index)
-                }
             }
         }
     }
