+++ feature/search/src/main/java/com/merxury/blocker/feature/search/screen/SearchResultScreen.kt
@@ -74,6 +74,7 @@ fun SearchResultScreen(
                 onUninstallClick = {},
                 onEnableClick = {},
                 onDisableClick = {},
+                onServiceStateUpdate = { _, _ -> },
             )
 
             1 -> ComponentSearchResultContent(
